app/requirements.txt
fastapi[standard]
uvicorn
supabase
argon2-cffi
dotenv